Alles werkt na de les van 18 november

+++ src/main/java/nl/novi/hello/service/BookService.java
@@ -1,12 +1,13 @@
 package nl.novi.hello.service;
 
+import nl.novi.hello.exception.BadRequestException;
 import nl.novi.hello.exception.RecordNotFoundException;
 import nl.novi.hello.model.Book;
 import nl.novi.hello.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -31,49 +32,66 @@ public class BookService {
             return optionalBook.get();
         }
         else {
-            // exception
             throw new RecordNotFoundException("ID does not exist!!!");
         }
 
     }
 
     public void deleteBook(int id) {
-        bookRepository.deleteById(id);
-    }
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+        }
+        else {
+            throw new RecordNotFoundException("ID does not exist!!!");
+        }
+}
 
     public int addBook(Book book) {
+        String isbn = book.getIsbn();
+        List<Book> books = (List<Book>)bookRepository.findAllByIsbn(isbn);
+        if (books.size() > 0) {
+            throw new BadRequestException("Isbn already exists!!!");
+        }
+
         Book newBook = bookRepository.save(book);
         return newBook.getId();
     }
 
     public void updateBook(int id, Book book) {
-        Book existingBook = bookRepository.findById(id).orElse(null);
+        Optional<Book> optionalBook = bookRepository.findById(id);
 
-        if (!book.getTitle().isEmpty()) {
-            existingBook.setTitle(book.getTitle());
+        if (optionalBook.isPresent()) {
+            Book storedBook = optionalBook.get();
+
+            book.setId(storedBook.getId());
+            bookRepository.save(book);
         }
-        if (!book.getAuthor().isEmpty()) {
-            existingBook.setAuthor(book.getAuthor());
+        else {
+            throw new RecordNotFoundException("ID does not exist!!!");
         }
-        if (!book.getIsbn().isEmpty()) {
-            existingBook.setIsbn(book.getIsbn());
-        }
-        bookRepository.save(existingBook);
     }
 
     public void partialUpdateBook(int id, Book book) {
-        Book existingBook = bookRepository.findById(id).orElse(null);
+        Optional<Book> optionalBook = bookRepository.findById(id);
 
-        if (!book.getTitle().isEmpty()) {
-            existingBook.setTitle(book.getTitle());
+        if (optionalBook.isPresent()) {
+            Book storedBook = bookRepository.findById(id).orElse(null);
+
+            if (book.getTitle()!=null && !book.getTitle().isEmpty()) {
+                storedBook.setTitle(book.getTitle());
+            }
+            if (book.getAuthor()!=null && !book.getAuthor().isEmpty()) {
+                storedBook.setAuthor(book.getAuthor());
+            }
+            if (book.getIsbn()!=null && !book.getIsbn().isEmpty()) {
+                storedBook.setIsbn(book.getIsbn());
+            }
+            bookRepository.save(storedBook);
+
         }
-        if (!book.getAuthor().isEmpty()) {
-            existingBook.setAuthor(book.getAuthor());
+        else {
+            throw new RecordNotFoundException("ID does not exist!!!");
         }
-        if (!book.getIsbn().isEmpty()) {
-            existingBook.setIsbn(book.getIsbn());
-        }
-        bookRepository.save(existingBook);
     }
 
 }
